Added unit tests for CompletionItems.
